+++ kernel/pic/pic.h
@@ -13,7 +13,7 @@
 void pic_remap(void); // Simple remap function
 void pic_disable(void); // Mask all interrupts
 uint16_t pic_get_mask(void); // Returns which IRQs are masked/unmasked
-void pic_send_eoi(uint8_t IRQ_line); // Helper function to send pic EOI acknowledgement
+void pic_send_eoi(uint32_t IRQ_line); // Helper function to send pic EOI acknowledgement
 void irq_enable(uint8_t IRQ_line); // Unmask the specific IRQ 
 void irq_disable(uint8_t IRQ_line); // Mask the specific IRQ
